requirements.txt
Django==4.0.5
djangorestframework==3.13.1
drf-yasg==1.20.0
psycopg2==2.9.3
python-dotenv==0.20.0

pytest-django==4.5.2
model_bakery==1.5.0
Faker==13.12.0
isort==5.10.1